update example of echo
