+++ service/grpcsrv/example/echo/server/server.go
@@ -1,17 +1,20 @@
 package main
 
 import (
+	"fmt"
+
+	"github.com/Sirupsen/logrus"
 	"github.com/lworkltd/kits/service/grpcsrv"
 	"github.com/lworkltd/kits/service/grpcsrv/example/echo/pb"
-	"github.com/lworkltd/kits/service/restful/code"
 )
 
 func main() {
+	logrus.SetFormatter(&logrus.TextFormatter{DisableColors: true})
 	grpcsrv.UseHook(grpcsrv.HookLogger)
 	grpcsrv.Register("Echo", func(req *pb.EchoRequest) (*pb.EchoResponse, error) {
 		return &pb.EchoResponse{
 			Str: req.Str,
-		}, code.New(1000, "")
+		}, fmt.Errorf("error")
 	})
 
 	grpcsrv.Run(":8090", "ECHO_SERVER")
